Refactor GraphQL structure: remove proposal.ts, update index.ts to export from degov.ts
